+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master (unreleased)
 
+## 0.37.0 (2023-06-20)
+
 * Add parsing `PivotTableDefinition#data_fields`
 * Add parsing `PivotTableDefinition#page_fields`
 * Add parsing `PivotTableDefinition#row_fields`